requirements.txt
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
httpx>=0.27.0
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
gitpython>=3.1.40
beautifulsoup4>=4.12.0
